add: TH fragments vente-card et recherche pour liste des ventes add: RechercheRecord pour ... la recherche De la refacto de code (service & repository) pour les Utilisateur et ArticleVendu

+++ src/main/java/fr/eni/projeteniencheres/dal/interfaces/ArticleVenduRepository.java
@@ -28,8 +28,8 @@ public interface ArticleVenduRepository {
 
     Integer terminerVente(ArticleVendu articleVendu);
 
-    public ArticleVendu findById(int id);
-    public List<ArticleVendu> findById(List<Integer> ids);
+    public ArticleVendu findById(long id);
+    public List<ArticleVendu> findById(List<Long> ids);
 
     public ArticleVendu save(ArticleVendu vente);
